+++ test/morgan.js
@@ -563,6 +563,63 @@ describe('morgan()', function () {
         .expect(200, cb)
       })
 
+      it('should have three digits by default', function (done) {
+        var cb = after(2, function (err, res, line) {
+          if (err) return done(err)
+          var end = Date.now()
+          assert.ok(/^[0-9]+\.[0-9]{3}$/.test(line))
+          done()
+        })
+
+        var stream = createLineStream(function (line) {
+          cb(null, null, line)
+        })
+
+        var start = Date.now()
+
+        request(createServer(':response-time', { stream: stream }))
+        .get('/')
+        .expect(200, cb)
+      })
+
+      it('should have five digits with argument "5"', function (done) {
+        var cb = after(2, function (err, res, line) {
+          if (err) return done(err)
+          var end = Date.now()
+          assert.ok(/^[0-9]+\.[0-9]{5}$/.test(line))
+          done()
+        })
+
+        var stream = createLineStream(function (line) {
+          cb(null, null, line)
+        })
+
+        var start = Date.now()
+
+        request(createServer(':response-time[5]', { stream: stream }))
+        .get('/')
+        .expect(200, cb)
+      })
+
+      it('should have no digits with argument "0"', function (done) {
+        var cb = after(2, function (err, res, line) {
+          if (err) return done(err)
+          var end = Date.now()
+          assert.ok(/^[0-9]+$/.test(line))
+          done()
+        })
+
+        var stream = createLineStream(function (line) {
+          cb(null, null, line)
+        })
+
+        var start = Date.now()
+
+        request(createServer(':response-time[0]', { stream: stream }))
+        .get('/')
+        .expect(200, cb)
+      })
+
       it('should not include response latency', function (done) {
         var cb = after(2, function (err, res, line) {
           if (err) return done(err)